Task 2, 7, 8 and 9
2) The current saving of a rate is carried out via a single POST to a Save API call, for
both inserts and updates. The API has now been amended to work with a POST and
a PUT independently. Please refactor the code to use these two calls. Please note,
the LengthOfStay value is not to be displayed and defaulted to 7, the Currency value
is readonly and defaulted to ‘EUR’.
7) When there are gaps in the rate dates provided, a warning message will be return
from the API call, this message should be displayed at the top of the list in a design
style which matches the rest of the application design. This message will be returned
with the rate data from a POST and PUT call. And also from the GET call when
listing the rates initially.
8) When inserting or updating a rate, the rate list should be refreshed to ensure all rates
are shown in correct date order.
9) Ensure that the view only version of this screen, as per the Agent user role
requirement, is inline with the above changes.